==> jupyterLab/OLD_Project_1_errorAnalysis/Step_1.2_Annot/1_Apr.24_Decompress_and_save_annot.ipynb <==
## sample (middle part)

ERR6684689.annot.vcf.gz   ERR7853477.annot.vcf.gz  SRR20975325.annot.vcf.gz  SRR23816670.annot.vcf.gz
ERR6684690.annot.vcf.gz   ERR7853480.annot.vcf.gz  SRR20975327.annot.vcf.gz  SRR23816671.annot.vcf.gz
ERR6684691.annot.vcf.gz   ERR7853482.annot.vcf.gz  SRR20975335.annot.vcf.gz  SRR23816672.annot.vcf.gz
ERR6684695.annot.vcf.gz   ERR7853484.annot.vcf.gz  SRR20975342.annot.vcf.gz  SRR23816673.annot.vcf.gz
ERR6684697.annot.vcf.gz   ERR7853485.annot.vcf.gz  SRR20975352.annot.vcf.gz  SRR23816674.annot.vcf.gz
ERR6684700.annot.vcf.gz   ERR7853489.annot.vcf.gz  SRR20975354.annot.vcf.gz  SRR23816675.annot.vcf.gz
ERR6684701.annot.vcf.gz   ERR7853491.annot.vcf.gz  SRR20975360.annot.vcf.gz  SRR23816677.annot.vcf.gz
ERR6684702.annot.vcf.gz   ERR7853492.annot.vcf.gz  SRR20975371.annot.vcf.gz  SRR23816679.annot.vcf.gz
ERR6684703.annot.vcf.gz   ERR7853493.annot.vcf.gz  SRR20975373.annot.vcf.gz  SRR23816680.annot.vcf.gz
ERR6684704.annot.vcf.gz   ERR7853497.annot.vcf.gz  SRR20975384.annot.vcf.gz  SRR23816681.annot.vcf.gz
ERR6684709.annot.vcf.gz   ERR7853500.annot.vcf.gz  SRR20975385.annot.vcf.gz  SRR23816682.annot.vcf.gz
ERR6684710.annot.vcf.gz   ERR7853502.annot.vcf.gz  SRR20975391.annot.vcf.gz  SRR23816683.annot.vcf.gz
ERR6684711.annot.vcf.gz   ERR7853504.annot.vcf.gz  SRR20975400.annot.vcf.gz  SRR23816684.annot.vcf.gz
ERR6684712.annot.vcf.gz   ERR7853512.annot.vcf.gz  SRR20975405.annot.vcf.gz  SRR23816685.annot.vcf.gz
ERR6684713.annot.vcf.gz   ERR7853514.annot.vcf.gz  SRR20975407.annot.vcf.gz  SRR23816686.annot.vcf.gz
ERR6684714.annot.vcf.gz   ERR7853515.annot.vcf.gz  SRR20975408.annot.vcf.gz  SRR23816688.annot.vcf.gz
ERR6684715.annot.vcf.gz   ERR7853516.annot.vcf.gz  SRR20975431.annot.vcf.gz  SRR23816690.annot.vcf.gz
ERR6684717.annot.vcf.gz   ERR7853517.annot.vcf.gz  SRR20975447.annot.vcf.gz  SRR23816691.annot.vcf.gz
ERR6684718.annot.vcf.gz   ERR7853519.annot.vcf.gz  SRR20975457.annot.vcf.gz  SRR23816692.annot.vcf.gz
ERR6684720.annot.vcf.gz   ERR7853527.annot.vcf.gz  SRR20975458.annot.vcf.gz  SRR23816693.annot.vcf.gz
ERR6684723.annot.vcf.gz   ERR7853528.annot.vcf.gz  SRR20975470.annot.vcf.gz  SRR23816694.annot.vcf.gz
ERR6684724.annot.vcf.gz   ERR7853532.annot.vcf.gz  SRR20975477.annot.vcf.gz  SRR23816695.annot.vcf.gz
ERR6684726.annot.vcf.gz   ERR7853535.annot.vcf.gz  SRR20975481.annot.vcf.gz  SRR23816696.annot.vcf.gz
ERR6684728.annot.vcf.gz   ERR7853542.annot.vcf.gz  SRR20975482.annot.vcf.gz  SRR23816697.annot.vcf.gz
ERR6684732.annot.vcf.gz   ERR7853543.annot.vcf.gz  SRR20975486.annot.vcf.gz  SRR23816698.annot.vcf.gz
ERR6684735.annot.vcf.gz   ERR7853546.annot.vcf.gz  SRR20975491.annot.vcf.gz  SRR23816700.annot.vcf.gz
ERR6684736.annot.vcf.gz   ERR7853550.annot.vcf.gz  SRR20975493.annot.vcf.gz  SRR23816701.annot.vcf.gz
ERR6684737.annot.vcf.gz   ERR7853551.annot.vcf.gz  SRR20975497.annot.vcf.gz  SRR23816702.annot.vcf.gz
ERR6684739.annot.vcf.gz   ERR7853554.annot.vcf.gz  SRR20975499.annot.vcf.gz  SRR23816703.annot.vcf.gz
ERR6684741.annot.vcf.gz   ERR7853556.annot.vcf.gz  SRR20975519.annot.vcf.gz  SRR23816704.annot.vcf.gz
ERR6684742.annot.vcf.gz   ERR7853559.annot.vcf.gz  SRR20975520.annot.vcf.gz  SRR23816705.annot.vcf.gz
ERR6684743.annot.vcf.gz   ERR7853560.annot.vcf.gz  SRR20975521.annot.vcf.gz  SRR23816706.annot.vcf.gz
ERR6684744.annot.vcf.gz   ERR7853561.annot.vcf.gz  SRR20975524.annot.vcf.gz  SRR23816707.annot.vcf.gz
ERR6684745.annot.vcf.gz   ERR7853562.annot.vcf.gz  SRR20975528.annot.vcf.gz  SRR23816708.annot.vcf.gz
ERR6684748.annot.vcf.gz   ERR7853565.annot.vcf.gz  SRR20975533.annot.vcf.gz  SRR23816711.annot.vcf.gz
ERR6684751.annot.vcf.gz   ERR7853566.annot.vcf.gz  SRR20975535.annot.vcf.gz  SRR23816712.annot.vcf.gz
ERR6684753.annot.vcf.gz   ERR7853568.annot.vcf.gz  SRR20975544.annot.vcf.gz  SRR23816713.annot.vcf.gz
ERR6684754.annot.vcf.gz   ERR7853569.annot.vcf.gz  SRR20975545.annot.vcf.gz  SRR23816714.annot.vcf.gz
ERR6684756.annot.vcf.gz   ERR7853572.annot.vcf.gz  SRR20975550.annot.vcf.gz  SRR23816715.annot.vcf.gz
ERR6684759.annot.vcf.gz   ERR7853576.annot.vcf.gz  SRR20975551.annot.vcf.gz  SRR23816716.annot.vcf.gz
ERR6684761.annot.vcf.gz   ERR7853577.annot.vcf.gz  SRR20975552.annot.vcf.gz  SRR23816717.annot.vcf.gz
ERR6684763.annot.vcf.gz   ERR7853581.annot.vcf.gz  SRR20975553.annot.vcf.gz  SRR23816718.annot.vcf.gz
ERR6684765.annot.vcf.gz   ERR7853582.annot.vcf.gz  SRR20975557.annot.vcf.gz  SRR23816719.annot.vcf.gz
ERR6684768.annot.vcf.gz   ERR7853585.annot.vcf.gz  SRR20975558.annot.vcf.gz  SRR23816720.annot.vcf.gz
ERR6684769.annot.vcf.gz   ERR7853587.annot.vcf.gz  SRR20975560.annot.vcf.gz  SRR23816721.annot.vcf.gz
ERR6684770.annot.vcf.gz   ERR7853590.annot.vcf.gz  SRR20975561.annot.vcf.gz  SRR23816722.annot.vcf.gz
ERR6684772.annot.vcf.gz   ERR7853592.annot.vcf.gz  SRR20975565.annot.vcf.gz  SRR23816723.annot.vcf.gz
ERR6684773.annot.vcf.gz   ERR7853595.annot.vcf.gz  SRR20975566.annot.vcf.gz  SRR23816724.annot.vcf.gz
ERR6684774.annot.vcf.gz   ERR7853598.annot.vcf.gz  SRR20975568.annot.vcf.gz  SRR23816725.annot.vcf.gz
ERR6684775.annot.vcf.gz   ERR7853600.annot.vcf.gz  SRR20975581.annot.vcf.gz  SRR23816726.annot.vcf.gz
ERR6684778.annot.vcf.gz   ERR7853601.annot.vcf.gz  SRR20975588.annot.vcf.gz  SRR23816727.annot.vcf.gz
ERR6684779.annot.vcf.gz   ERR7853602.annot.vcf.gz  SRR20975589.annot.vcf.gz  SRR23816728.annot.vcf.gz
ERR6684780.annot.vcf.gz   ERR7853603.annot.vcf.gz  SRR20975592.annot.vcf.gz  SRR23816729.annot.vcf.gz
ERR6684781.annot.vcf.gz   ERR7853604.annot.vcf.gz  SRR20975594.annot.vcf.gz  SRR23816730.annot.vcf.gz
ERR6684782.annot.vcf.gz   ERR7853608.annot.vcf.gz  SRR20975595.annot.vcf.gz  SRR23816731.annot.vcf.gz
ERR6684783.annot.vcf.gz   ERR7853610.annot.vcf.gz  SRR20975613.annot.vcf.gz  SRR23816732.annot.vcf.gz
ERR6684784.annot.vcf.gz   ERR7853613.annot.vcf.gz  SRR20975618.annot.vcf.gz  SRR23816733.annot.vcf.gz
ERR6684786.annot.vcf.gz   ERR7853614.annot.vcf.gz  SRR20975621.annot.vcf.gz  SRR23816734.annot.vcf.gz
ERR6684787.annot.vcf.gz   ERR7853615.annot.vcf.gz  SRR20975627.annot.vcf.gz  SRR23816735.annot.vcf.gz
ERR6684788.annot.vcf.gz   ERR7853616.annot.vcf.gz  SRR20975631.annot.vcf.gz  SRR23816736.annot.vcf.gz
ERR6684790.annot.vcf.gz   ERR7853617.annot.vcf.gz  SRR20975632.annot.vcf.gz  SRR23816737.annot.vcf.gz
ERR6684791.annot.vcf.gz   ERR7853620.annot.vcf.gz  SRR20975634.annot.vcf.gz  SRR23816738.annot.vcf.gz
ERR6684792.annot.vcf.gz   ERR7853621.annot.vcf.gz  SRR20975636.annot.vcf.gz  SRR23816739.annot.vcf.gz
ERR6684794.annot.vcf.gz   ERR7853622.annot.vcf.gz  SRR20975641.annot.vcf.gz
ERR6684798.annot.vcf.gz   ERR7853623.annot.vcf.gz  SRR20975647.annot.vcf.gz

import pandas as pd
import numpy as np

# Define function: split each key-value pair in INFO into two columns
def parse_info_field(info_str):
    fields = info_str.split(';')
    keys = []
    values = []
    for field in fields:
        if '=' in field:
            key, value = field.split('=')
            keys.append(key)
            values.append(value)
        else:
            keys.append(field)
            values.append(True)
    return pd.Series(values, index=keys)

# Read vcf file
df = pd.read_csv('/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Downloads/SRR20975566.annot.vcf.gz', delimiter='\t',comment='#', header=None,
                 dtype={0: str, 1: np.int64, 2: str, 3: str, 4: str, 5: str, 6: str, 7: str}, compression='gzip')

# Set column names
df.columns = ['#CHROM', 'POS', 'ID', 'REF', 'ALT', 'QUAL', 'FILTER', 'INFO']

# Split the INFO column into multiple key-value pairs
df_info = df['INFO'].apply(parse_info_field)

# Add the processed result to the original data frame
df = pd.concat([df, df_info], axis=1)

# Drop rows based on the value of the 'INDEL' column, if it is present
if 'INDEL' in df.columns:
    df = df.drop(df[df['INDEL'] == True].index)

# Create AF and SB columns and set them to 0 if they don't exist
if 'AF' not in df.columns:
    df['AF'] = 0
if 'SB' not in df.columns:
    df['SB'] = 0

# Transform AF > 0.5
df['AF'] = df['AF'].astype(float)
df.loc[df['AF'] > 0.5, 'AF'] = 1 - df['AF']

#### >>>>>>>>>>>>>>>>>> del eg:AT->T <<<<<<<<<<<<<<<<<
# Ensure 'REF' and 'ALT' are of type string
df['REF'] = df['REF'].astype(str)
df['ALT'] = df['ALT'].astype(str)

# Filter rows where 'REF' and 'ALT' have only one letter
df = df[(df['REF'].str.len() <= 1)]
#### >>>>>>>>>>>>>>>>>> del eg:AT->T <<<<<<<<<<<<<<<<<

# Extract desired columns
df = df[['POS', 'REF', 'ALT', 'AF', 'SB']]

# Drop duplicates in 'POS' column
df = df.drop_duplicates(subset='POS', keep='first')

# Create a new index
new_index = pd.RangeIndex(start=1, stop=29904, step=1)

# Re-index the data frame
df = df.set_index('POS').reindex(new_index, fill_value=0).reset_index(drop=False).rename(columns={'index': 'POS'})

df[df['SB'] !=0 ]

import pandas as pd
import numpy as np

# Define function: split each key-value pair in INFO into two columns
def parse_info_field(info_str):
    fields = info_str.split(';')
    keys = []
    values = []
    for field in fields:
        if '=' in field:
            key, value = field.split('=')
            keys.append(key)
            values.append(value)
        else:
            keys.append(field)
            values.append(True)
    return pd.Series(values, index=keys)

# Read vcf file
df = pd.read_csv('/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Downloads/SRR23816718.annot.vcf.gz', delimiter='\t',comment='#', header=None,
                 dtype={0: str, 1: np.int64, 2: str, 3: str, 4: str, 5: str, 6: str, 7: str}, compression='gzip')

# Set column names
df.columns = ['#CHROM', 'POS', 'ID', 'REF', 'ALT', 'QUAL', 'FILTER', 'INFO']

# Split the INFO column into multiple key-value pairs
df_info = df['INFO'].apply(parse_info_field)

# Add the processed result to the original data frame
df = pd.concat([df, df_info], axis=1)

# Drop rows based on the value of the 'INDEL' column, if it is present
if 'INDEL' in df.columns:
    df = df.drop(df[df['INDEL'] == True].index)

# Create AF and SB columns and set them to 0 if they don't exist
if 'AF' not in df.columns:
    df['AF'] = 0
if 'SB' not in df.columns:
    df['SB'] = 0

# Transform AF > 0.5
df['AF'] = df['AF'].astype(float)
df.loc[df['AF'] > 0.5, 'AF'] = 1 - df['AF']

#### >>>>>>>>>>>>>>>>>> del eg:AT->T <<<<<<<<<<<<<<<<<
# Ensure 'REF' and 'ALT' are of type string
df['REF'] = df['REF'].astype(str)
df['ALT'] = df['ALT'].astype(str)

# Filter rows where 'REF' and 'ALT' have only one letter
df = df[(df['REF'].str.len() <= 1) & (df['ALT'].str.len() <= 1)]
#### >>>>>>>>>>>>>>>>>> del eg:AT->T <<<<<<<<<<<<<<<<<

# Extract desired columns
df = df[['POS', 'REF', 'ALT', 'AF', 'SB']]

# Drop duplicates in 'POS' column
df = df.drop_duplicates(subset='POS', keep='first')

# Create a new index
new_index = pd.RangeIndex(start=1, stop=29904, step=1)

# Re-index the data frame
df = df.set_index('POS').reindex(new_index, fill_value=0).reset_index(drop=False).rename(columns={'index': 'POS'})

df

## For test / also for final version

##### Code block:
```bash
bsub -M 2000 
-e /nfs/research/goldman/zihao/errorsProject_1/Annot/Annot_decompress_errorChecking_error.txt 
'python3 /nfs/research/goldman/zihao/errorsProject_1/Annot/Annot_Decompress_and_save.py'
```

##### csv版本_test_封装class后，未封装版在Annot_Decompress_and_save.py'

import csv
import gzip
import glob
import os

class VCFProcessor:
    def __init__(self, input_dir, output_dir):
        self.input_dir = input_dir
        self.output_dir = output_dir

    @staticmethod
    def parse_info_field(info_str):
        """Parse the INFO field of a VCF file and return a dictionary."""
        info_dict = {}
        for field in info_str.split(';'):
            if '=' in field:
                key, value = field.split('=')
                info_dict[key] = value
        return info_dict

    def process_vcf_file(self, file_path):
        """Process a single VCF file and return the processed data."""
        try:
            data = {}
            new_data = []

            with gzip.open(file_path, 'rt') as file:
                reader = csv.reader(file, delimiter='\t')

                for row in reader:
                    if not row or row[0].startswith('#'):
                        continue

                    chrom, pos, id_, ref, alt, qual, filter_, info = row
                    pos = int(pos)
                    info_dict = self.parse_info_field(info)

                    if 'INDEL' in info_dict and info_dict['INDEL'] == True:
                        continue
                        
                    # Skip the current loop if 'REF' or 'ALT' have more than one character
                    if len(ref) > 1 or len(alt) > 1:
                        continue

                    ref = ref if ref else '0'
                    alt = alt if alt else '0'
                    af = info_dict.get('AF', '0')
                    sb = info_dict.get('SB', '0')
                    
                    af = float(af)
                    if af > 0.5:
                        af = 1 - af

                    if pos not in data:
                        data[pos] = [pos, ref, alt, af, sb]

            for idx in range(1, length_of_sample):
                if idx in data:
                    new_data.append(data[idx])
                else:
                    new_data.append([idx, 'NA', 'NA', '0', '0'])

            return new_data

        except Exception as e:
            print(f"Unknown error processing {file_path}: {e}")

    def process_files(self):
        """Process all VCF files in the input directory and write the results to the output directory."""
        
        file_paths = glob.glob(os.path.join(self.input_dir, '*.annot.vcf.gz'))

        for i, file_path in enumerate(file_paths):
            
            # Remove this line for the final version！！！！
            if i >= 10:  
                break
            # Remove this line for the final version！！！！

            output_file = os.path.join(self.output_dir, os.path.basename(file_path).replace('.annot.vcf.gz', '_annot.txt'))

            if os.path.exists(output_file):
                print(f"{output_file} already exists. Skipping file {file_path}.")
                continue

            result = self.process_vcf_file(file_path)
            if result is None:
                continue

            with open(output_file, 'w', newline='') as f:
                writer = csv.writer(f, delimiter='\t')
                writer.writerow(['POS', 'REF', 'ALT', 'AF', 'SB'])
                writer.writerows(result)

for i in range(1, 3):
    print(i)

if __name__ == '__main__':
    length_of_sample = 29904 # the true number +1 (eg: 29903 should be 29904)
    input_dir = '/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Downloads/'
    output_dir = '/homes/zihao/DATAS/TEST_for_annot_new_may/'
    processor = VCFProcessor(input_dir, output_dir)
    processor.process_files()
