==> jupyterLab/OLD_Project_1_errorAnalysis/Step_1.2_Annot/2.1_Apr.24_Data Processing_for_error_pos.ipynb <==
```bash
bsub -M 2000 -e /nfs/research/goldman/zihao/errorsProject_1/Annot/Treat_error_pos_errorChecking_error.txt 'python3 /nfs/research/goldman/zihao/errorsProject_1/Annot/Annot_Treat_error_pos.py'
```
```python
input: "/nfs/research/goldman/zihao/errorsProject_1/MAPLE/TEST_50000/output_modified.txt"
output: "/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Error_pos_for_annot.txt" 
```

# Data Processing

df.head()

##### Rename (!!! Just for test)

import pandas as pd
import os

df = pd.read_csv("/homes/zihao/EBI_INTER/P1_Coverage/output_modified.txt", sep='\t')

folder_path = '/homes/zihao/DATAS/TEST_for_annot/'
file_extension = '_annot.txt'
file_list = os.listdir(folder_path)
id_list = df['ID'].to_list()

for i in range(len(file_list)):
    old_file_name = file_list[i]
    new_file_name = id_list[i] + file_extension
    os.rename(os.path.join(folder_path, old_file_name), os.path.join(folder_path, new_file_name))


## 1. handle with the datas (For test also for final version)_output.txt

- **Input**:output_modified.txt
- **Output**:processed_data_Annot.txt [output.txt (old)]

### old version
```python
# 读取数据1
data1_file = "/homes/zihao/EBI_INTER/P1_Coverage/output_modified.txt"
data1 = []
with open(data1_file) as f:
    header = f.readline().strip() + "\tAF\tSB\n"
    for line in f:
        line = line.strip().split()
        data1.append(line)

# 遍历数据1中的ID，并在文件名中寻找含有该ID的文件
data_folder = "/homes/zihao/DATAS/TEST_for_annot/"
for row in data1:
    id = row[0]
    filename = os.path.join(data_folder, f"{id}_annot.txt")
    if os.path.exists(filename):
        # 打开含有该ID的文件
        with open(filename) as f:
            for line in f:
                line = line.strip().split()
                # 如果POS列与数据1中第二列相同，则将该行信息添加到数据1新的列中
                if line[0] == row[1]:
                    row.append(line[3])
                    row.append(line[4])
                    break
            else:
                # 如果未找到匹配行，则在该行中添加空值
                row.append("")
                row.append("")
    else:
        # 如果文件不存在，则在该行中添加空值
        row.append("")
        row.append("")

# 将数据1输出到文件中
output_file = "output.txt"
with open(output_file, "w") as f:
    f.write(header)
    for row in data1:
        f.write("\t".join(row) + "\n")
```

import os

class DataProcessor:
    def __init__(self, data1_file, data_folder, output_file):
        self.data1_file = data1_file
        self.data_folder = data_folder
        self.output_file = output_file

    def read_data1(self):
        """Read data1 from file and return the header and data."""
        data1 = []
        with open(self.data1_file) as f:
            header = f.readline().strip() + "\tAF\tSB\n"
            for line in f:
                line = line.strip().split()
                data1.append(line)
        return header, data1

    def process_data(self):
        """Process data1 by looking up corresponding rows in data folder."""
        header, data1 = self.read_data1()

        for row in data1:
            id = row[0]
            filename = os.path.join(self.data_folder, f"{id}_annot.txt")
            if os.path.exists(filename):
                with open(filename) as f:
                    for line in f:
                        line = line.strip().split()
                        if line[0] == row[1]:
                            row.append(line[3])
                            row.append(line[4])
                            break
                    else:
                        row.append("")
                        row.append("")
            else:
                row.append("")
                row.append("")

        return header, data1

    def write_output(self):
        """Write the processed data1 to the output file."""
        header, data1 = self.process_data()

        with open(self.output_file, "w") as f:
            f.write(header)
            for row in data1:
                f.write("\t".join(row) + "\n")

if __name__ == '__main__':
    data1_file = "/nfs/research/goldman/zihao/errorsProject_1/MAPLE/TEST_50000/output_modified.txt"
    data_folder = "/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Decompress/"
    output_file = "/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Error_pos_for_annot.txt" # 可修改添加输出文件夹

    processor = DataProcessor(data1_file, data_folder, output_file)
    processor.write_output()
