==> jupyterLab/OLD_Project_1_errorAnalysis/Step_1.2_Annot/3_Final_plot_Annot-May.ipynb <==
import plotly.graph_objects as go
from plotly.offline import iplot, plot
import pandas as pd
import numpy as np

import plotly
from plotly.subplots import make_subplots
import plotly.io as pio
from scipy.stats import gaussian_kde

### 1. For all

## For all
# df_all = pd.read_csv('../../selected_data.txt', sep='\t')
df_all = pd.read_csv('/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Annot_RATIO_sampling_all.txt', sep='\t')
df_all['AF'] = df_all['AF'].replace([np.inf, -np.inf], np.nan)  # 将无穷大和无穷小值替换为NaN
df_all = df_all.dropna()  # 删除包含NaN的行
df_all

### 2. For error_pos

# df_err = pd.read_csv("../../A_Datas/Error_pos_for_annot.txt", sep='\t')
df_err = pd.read_csv("/nfs/research/goldman/zihao/Datas/p1/File_5_annot/Error_pos_for_annot_all.txt", sep='\t')
df_err['AF'] = df_err['AF'].replace([np.inf, -np.inf], np.nan)  # 将无穷大和无穷小值替换为NaN
df_err = df_err.dropna()  # 删除包含NaN的行
df_err

### Final version！！！

import plotly
import plotly.graph_objects as go
from plotly.subplots import make_subplots
import plotly.io as pio
from scipy.stats import gaussian_kde

hist_data = [(df_err[df_err['AF'] * 100 < 1]['AF'] * 100).tolist(), 
             (df_all[df_all['AF'] * 100 < 1]['AF'] * 100).tolist()]
group_labels = ['Error Position', 'All Position'] # name of the dataset
colors = ['#835AF1', '#B8F7D4']

fig = make_subplots(rows=2, cols=1, row_heights=[0.75, 0.25], shared_xaxes=True)

# 添加Density Plot子图
for data, label, color in zip(hist_data, group_labels, colors):
    kde = gaussian_kde(data)
    x_vals = np.linspace(min(data), max(data), 1000)
    y_vals = kde(x_vals)
    
    fig.add_trace(go.Scatter(x=x_vals, y=y_vals, mode='lines', fill='tozeroy', showlegend=False, line_color=color), 
                  row=1, col=1)

# 添加Box Plot子图
fig.add_trace(go.Box(x=df_all['AF'].tolist(), orientation='h', name=group_labels[1]), row=2, col=1)
fig.add_trace(go.Box(x=df_err['AF'].tolist(), orientation='h', name=group_labels[0]), row=2, col=1)

# 设置图形布局
fig.update_layout(
    title="Allele Frequency density map over sample positions",
    yaxis=dict(
        title='Probability density',
        showline=True, showgrid=False,
        linewidth=2, linecolor='gray', ticks='outside',
        tickfont=dict(
            family='Arial',
            size=12,
            color='black',
        )
    ),
    xaxis=dict(
        title='Allele Frequency / %',
        # type='log',  # 对数刻度
        range=[0, 0.45],
        showline=True, showgrid=False,
        linewidth=2, linecolor='gray', ticks='outside',
        tickfont=dict(
            family='Arial',
            size=12,
            color='black',
        )
    ),
    legend=dict(
        traceorder="normal",
        font=dict(size=12),
    ),
    plot_bgcolor='white',
    yaxis_gridcolor='lightgray', yaxis_gridwidth=0.5,
    xaxis_gridcolor='lightgray', xaxis_gridwidth=0.5,
)

#fig.show()
pio.write_image(fig, 'May.18_plot_for_AF_n.png')
#plotly.offline.plot(fig, filename = "plot_for_AF_nn.html")

### rug plot：
Showing the original discrete distribution of the data

import gc
gc.collect()
